Implement mediator pattern and WIP cache
