+++ lib/config.ts
@@ -14,6 +14,7 @@ export type IgnoreUsers = {
   irc?: string[];
   discord?: string[];
   discordIds?: string[];
+  roles?: string[];
 };
 
 export type GameLogConfig = {
@@ -78,6 +79,7 @@ export const IgnoreUsersSchema = z.object({
   irc: z.array(z.string()).optional(),
   discord: z.array(z.string()).optional(),
   discordIds: z.array(z.string()).optional(),
+  roles: z.array(z.string()).optional(),
 });
 
 export const GameLogConfigSchema = z.object({
